requirements_test.txt
-r requirements.txt
bandit
black
build
bumpver
isort
pip-tools
pre-commit
pylint
pytest
pytest-asyncio
python-dotenv
reorder-python-imports
twine
